test(sandbox): fold base public API contract tests into test_sandbox.py
Keep a single provider-agnostic core sandbox test module instead of a
separate test_sandbox_base.py: the contract tests for the public types
(exports, SandboxStatus, SandboxResources, SandboxSpec, exec result,
handle, error hierarchy) now live next to the existing validation tests,
the builtin-loader test sits with the registry tests, and the config
resolution edge cases join test_resolve_provider_config_errors.

The spec-level resource coercion assertions were dropped from the new
tests since test_sandbox_resources_validation already covers them.
